processes: add support for most processes needed for a test system
in order to get a functional test system based on ZPC we depend on:
- commander-cli (to handle the interaction with the WPK/radio board attached)
- mosquitto (ZPC and unify tools need a MQTT broker)
- mosquitto_sub (to dump the unify mqtt messages in a file)
- socat (to get rid of the usb connection between ncp_controller and gateway)
- uic-upvl (for smartstart provisioning)
- uic-image-updater (for OTA)
- zpc (because we need a gateway, it allows us to do OTW updates as well)

apart from commander-cli, all of the executables above need to run in the
background, that's why they all derive from the BackgroundProcess class
which handles the lifecycle of the background process.

background_process_list is a module accessible list that contains references
to all the background processes at a given time. This allows fixtures to be
created to remove them on setup/teardown.