devices: add DevWpk and ZwaveDevBase classes
DevWpk abstracts a Silicon Labs WPK, more specifically its admin interface
exposed on port 4902. It allows the test system to retrieve information
about the radio board connected to the WPK.

ZwaveDevBase is a class that should contain common stuff between ncp and soc
apps. It needs a DevWpk and is used to flash a radio board with the right
firmware.