examples: add example test script using Pytest
also adds Pytest fixtures for handling background processes, logs and
the retrieval of DevWpk objects from the json description in the clusters.json
file.

added some basic log config for Pytest.
added a bash script to start the test script with the right options and
to move the pytest.log and junit-results.xml file into the test run folder.