arch: z_wave_ts_silabs is now a pytest plugin
make z_wave_ts_silabs a pytest plugin, this requires modifications
in the setup.py and fixtures.py files (formerly pytest_fixtures.py),
this allows us to remove conftest.py.

fixture.py will be used as an entrypoint. it now adds a new parser option
to the pytest commandline (--hw-cluster) that enables a user to provide
the identifier of the cluster the tests will run on.

run_test.sh has been modified to accommodate this change.

The requirements.txt now pulls the next branch of the package on github.
In an another project this should be a tag. We may change that later.

The README.md was updated with new commands to set up and update the
package locally.

the .gitignore was updated with folders created by pip wheel.

The handling of both config.json and clusters.json is now done in
config.py, as part of the init of the shared test system Context: ctxt,
for that purpose the from_json function of the Cluster class (clusters.py)
was changed to from_dict().

clusters was renamed to clusters_json in config.json, this was done to
free up the name clusters in ctxt so that it could be used to store a
list of Cluster(s).

We now use Path from pathlib as well. we should probably update the rest
of the code base later as well.

The example test_script uses the new fixtures and parameters to handle
the region and how to get their wpks, the fixture to get these wpks is
now scoped to 'session' instead of 'function'