refactor!(processes): move start out of init
The constructor of the class BackgroundProcess included logic, including
file operations and create threads. This functionality is now inside the
method start. Now, the client is responsible for starting the process
after the instance is created. Whereas before the processes was started
by the constructor.

This refactor is part of an effort to move code logic from the
constructors to dedicated functions, with the intend to increase the
granularity of the operations to enable new tests. e.g., tests that
require to restart processes.