cli: get_node_id command in the cli now return a 4 digit hex value
get_node_id command in the cli now return a 4 digit hex value, this is
to follow the same convention as ZPC even though the maximum node ID
value support in long range is 0xFFF, three digits would have been enough.

At least this way we can use the same regex to parse the output, it's
better in terms of consistency.