processes: fix zpc region configuration and use LR first if region is LR
fix zpc region configuration and use LR first if region is LR, this is
done through the protocol_pref option in uic.cfg, 1 is Z-Wave Classic,
2 is Z-Wave Long Range.

zpc region configuration only uses the abbreviation of the region without
the REGION_ prefix, if it's passed to _generate_uic_configuration_file
it's removed automatically.

also when using SmartStart the "PreferredProtocols" only need to be
specified if the user wants to override the protocol_pref in uic.cfg,
we now leave it blank so that it does not clash with uic.cfg.