parsers: added a zlf parser that returns the DCH packets from a file
also added a pcap parser that returns the raw Z-Wave frame from a file
with some metadata. The Z-Wave pcap definitions where move to the end
of the definitions.py module. A script was created to extract a csv from
a pcap file as a proof of concept for the PcapFileReader class.