SWPROT-9436: Fix inconsistent command output in CLI
Observed a bug where the output of a command was buffered and returned
as output of the next command.
Example:
2025-04-08 11:32:04.023 DevZwaveDoorLockKeypad-1 DEBUG _run_cmd: get_dsk
2025-04-08 11:32:05.025 DevZwaveDoorLockKeypad-1 DEBUG _run_cmd output: get_dsk
2025-04-08 11:32:05.026 DevZwaveDoorLockKeypad-1 DEBUG _run_cmd: set_learn_mode
2025-04-08 11:32:05.102 DevZwaveDoorLockKeypad-1 DEBUG _run_cmd output:

[I] 42032-00370-02109-57549-46761-28345-59600-29460

>

This might be because read_until method is based on
read_lazy and doesn't have command output in buffer at the time of
reading.
This commit aim to fix this bug by splitting the read in 2 steps. In
case the read isn't successful, we try to recover with eager read.