SWPROT-9631: Ignore invalid CLI characters at startup
Sometimes invalid characters are received on the CLI, expecially if the previous firmware was a FLiRs. To avoid crashing the test, ignore ASCII errors when spwaning a new device.
Note that the function read_very_eager() should flush the input buffer and avoid this kind of problems. But for unknow reason, we still have the error when calling _run_cmd.